pyproject.toml
[project]
name = "ramen-logistics-app"
version = "0.1.0"
description = "ラーメンチェーン店食材自動発注システム - Streamlit in Snowflake (SiS)"
requires-python = ">=3.9"
dependencies = [
    "snowflake-snowpark-python>=1.11.0,<2.0",
    "streamlit>=1.31.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "SIM"]
ignore = ["S101"]  # assert in tests is OK

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: 単体テスト（ローカル実行、モック使用）",
    "integration: 結合テスト（Snowflakeステージング環境）",
    "regression: 回帰テスト（Phase間移行時に実行）",
]
addopts = "--strict-markers"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
